Extract database bootstrap into container_bootstrap.rs
Step 4 of splitting containers.rs. Moves compose_exec,
postgres_database_exists_query, database_charset, and ensure_database
(readiness polling + create-database-if-missing for
Postgres/MySQL/MariaDB). database_charset is re-exported (`pub(crate)
use crate::container_bootstrap::database_charset;`) since backups.rs
calls it directly as containers::database_charset; ensure_database is
only called internally (operate_inner) so just imported. Its 4 tests
(postgres query builder + the 3 database_charset tests) moved with
it. 2201 -> 1945 lines (container_bootstrap.rs: 348).
