+++ src-tauri/src/containers.rs
@@ -9,6 +9,8 @@ use std::{
 };
 use tauri::Manager;
 
+pub(crate) use crate::container_bootstrap::database_charset;
+use crate::container_bootstrap::ensure_database;
 use crate::container_compose::{compose, php_dockerfile, site_document_root, site_hostnames};
 use crate::container_inspection::{
     apply_stats as apply_service_stats, parse_services as parse_service_inspections,
@@ -406,228 +408,6 @@ pub fn execute_service_command(
     } else {
         combined
     })
-}
-
-fn compose_exec(
-    executable: &str,
-    directory: &Path,
-    args: &[String],
-) -> Result<std::process::Output, String> {
-    crate::process::output(
-        runtime_command(executable)
-            .args(args)
-            .current_dir(directory)
-            .stdin(Stdio::null()),
-        crate::process::DATABASE_TIMEOUT,
-        "database container command",
-    )
-}
-
-fn postgres_database_exists_query(database_name: &str) -> String {
-    format!("SELECT 1 FROM pg_database WHERE datname = '{database_name}';")
-}
-
-/// Validated `DB_CHARSET` environment variable (set by the project wizard's
-/// "Encoding" field), or MySQL/MariaDB's sane default. Shared with
-/// `backups::clear_database_sql` so charset selection is consistent between
-/// initial creation and a manual "clear database".
-pub(crate) fn database_charset(environment: &Environment) -> &str {
-    environment
-        .environment_variables
-        .get("DB_CHARSET")
-        .filter(|value| {
-            !value.is_empty()
-                && value
-                    .chars()
-                    .all(|character| character.is_ascii_alphanumeric() || character == '_')
-        })
-        .map(String::as_str)
-        .unwrap_or("utf8mb4")
-}
-
-fn ensure_database(
-    app: &tauri::AppHandle,
-    id: &str,
-    executable: &str,
-    directory: &Path,
-    environment: &Environment,
-) -> Result<(), String> {
-    let mut ready = false;
-    let mut last_error = String::new();
-    // A fresh MySQL/MariaDB data directory can take considerably longer than
-    // 30 seconds to initialise on slower disks or immediately after an image pull.
-    for attempt in 0..60 {
-        let args = if environment.database == "PostgreSQL" {
-            vec![
-                "compose".into(),
-                "exec".into(),
-                "-T".into(),
-                "database".into(),
-                "pg_isready".into(),
-                "-U".into(),
-                environment.database_user.clone(),
-            ]
-        } else {
-            let admin = if environment.database == "MariaDB" {
-                "mariadb-admin"
-            } else {
-                "mysqladmin"
-            };
-            vec![
-                "compose".into(),
-                "exec".into(),
-                "-T".into(),
-                "database".into(),
-                admin.into(),
-                "ping".into(),
-                "-h".into(),
-                "127.0.0.1".into(),
-                "-uroot".into(),
-                format!("-p{}", environment.database_root_password),
-                "--silent".into(),
-            ]
-        };
-        match compose_exec(executable, directory, &args) {
-            Ok(output) if output.status.success() => {
-                ready = true;
-                break;
-            }
-            Ok(output) => {
-                let stderr = String::from_utf8_lossy(&output.stderr).trim().to_owned();
-                let stdout = String::from_utf8_lossy(&output.stdout).trim().to_owned();
-                last_error = if stderr.is_empty() { stdout } else { stderr };
-            }
-            Err(error) => last_error = error,
-        }
-        if attempt % 5 == 0 {
-            emit_progress(
-                app,
-                id,
-                85,
-                &format!(
-                    "Waiting for database initialization ({}s)",
-                    (attempt + 1) * 2
-                ),
-            );
-        }
-        thread::sleep(Duration::from_secs(2));
-    }
-    if !ready {
-        let log_args = vec![
-            "compose".into(),
-            "logs".into(),
-            "--no-color".into(),
-            "--tail".into(),
-            "40".into(),
-            "database".into(),
-        ];
-        let logs = compose_exec(executable, directory, &log_args)
-            .ok()
-            .map(|output| {
-                let stdout = String::from_utf8_lossy(&output.stdout).trim().to_owned();
-                let stderr = String::from_utf8_lossy(&output.stderr).trim().to_owned();
-                if stdout.is_empty() {
-                    stderr
-                } else {
-                    stdout
-                }
-            })
-            .unwrap_or_default();
-        let detail = if !logs.is_empty() {
-            logs
-        } else if !last_error.is_empty() {
-            last_error
-        } else {
-            "No database logs were returned".into()
-        };
-        return Err(format!(
-            "Database service did not become ready within 120 seconds.\n{detail}"
-        ));
-    }
-    // The project wizard's "Automatically create database" toggle is stored
-    // as an environment variable rather than a dedicated field; honor it
-    // here instead of always creating the database regardless of the user's
-    // choice. The database *server* still needs to be up first (waited for
-    // above) even when skipping creation, since an app installer or an
-    // imported SQL dump may create its own database.
-    if environment
-        .environment_variables
-        .get("LS_PANEL_AUTO_CREATE_DATABASE")
-        .map(String::as_str)
-        == Some("false")
-    {
-        return Ok(());
-    }
-    if environment.database == "PostgreSQL" {
-        let query = postgres_database_exists_query(&environment.database_name);
-        let query_args = vec![
-            "compose".into(),
-            "exec".into(),
-            "-T".into(),
-            "-e".into(),
-            format!("PGPASSWORD={}", environment.database_password),
-            "database".into(),
-            "psql".into(),
-            "-U".into(),
-            environment.database_user.clone(),
-            "-d".into(),
-            "postgres".into(),
-            "-tAc".into(),
-            query,
-        ];
-        let query_output = compose_exec(executable, directory, &query_args)?;
-        if !query_output.status.success() {
-            return Err(String::from_utf8_lossy(&query_output.stderr)
-                .trim()
-                .to_owned());
-        }
-        if String::from_utf8_lossy(&query_output.stdout).trim() == "1" {
-            return Ok(());
-        }
-        let create_args = vec![
-            "compose".into(),
-            "exec".into(),
-            "-T".into(),
-            "-e".into(),
-            format!("PGPASSWORD={}", environment.database_password),
-            "database".into(),
-            "createdb".into(),
-            "-U".into(),
-            environment.database_user.clone(),
-            environment.database_name.clone(),
-        ];
-        let output = compose_exec(executable, directory, &create_args)?;
-        return if output.status.success() {
-            Ok(())
-        } else {
-            Err(String::from_utf8_lossy(&output.stderr).trim().to_owned())
-        };
-    }
-
-    let client = if environment.database == "MariaDB" {
-        "mariadb"
-    } else {
-        "mysql"
-    };
-    let charset = database_charset(environment);
-    let sql = format!("CREATE DATABASE IF NOT EXISTS `{}` CHARACTER SET {charset}; CREATE USER IF NOT EXISTS '{}'@'%' IDENTIFIED BY '{}'; ALTER USER '{}'@'%' IDENTIFIED BY '{}'; GRANT ALL PRIVILEGES ON `{}`.* TO '{}'@'%'; FLUSH PRIVILEGES;", environment.database_name, environment.database_user, environment.database_password, environment.database_user, environment.database_password, environment.database_name, environment.database_user);
-    let args = vec![
-        "compose".into(),
-        "exec".into(),
-        "-T".into(),
-        "database".into(),
-        client.into(),
-        "-uroot".into(),
-        format!("-p{}", environment.database_root_password),
-        "-e".into(),
-        sql,
-    ];
-    let output = compose_exec(executable, directory, &args)?;
-    if output.status.success() {
-        Ok(())
-    } else {
-        Err(String::from_utf8_lossy(&output.stderr).trim().to_owned())
-    }
 }
 
 fn ensure_network(executable: &str) -> Result<(), String> {
@@ -1774,13 +1554,6 @@ mod tests {
     }
 
     #[test]
-    fn postgres_database_check_uses_server_sql_without_psql_meta_commands() {
-        let query = postgres_database_exists_query("app");
-        assert_eq!(query, "SELECT 1 FROM pg_database WHERE datname = 'app';");
-        assert!(!query.contains("\\gexec"));
-    }
-
-    #[test]
     fn local_route_accepts_application_404_but_not_server_errors() {
         assert!(route_status_is_available(200));
         assert!(route_status_is_available(404));
@@ -2168,34 +1941,5 @@ mod tests {
         assert!(!is_network_already_exists_error(
             "no such file or directory"
         ));
-    }
-
-    #[test]
-    fn database_charset_defaults_to_utf8mb4() {
-        let mut environment = test_environment();
-        environment.environment_variables.remove("DB_CHARSET");
-        assert_eq!(database_charset(&environment), "utf8mb4");
-    }
-
-    #[test]
-    fn database_charset_uses_a_valid_wizard_selection() {
-        let mut environment = test_environment();
-        environment
-            .environment_variables
-            .insert("DB_CHARSET".into(), "utf8".into());
-        assert_eq!(database_charset(&environment), "utf8");
-    }
-
-    #[test]
-    fn database_charset_rejects_values_that_are_not_a_safe_identifier() {
-        // A charset name is interpolated directly into `CREATE DATABASE ...
-        // CHARACTER SET {charset}` with no further escaping — reject
-        // anything that isn't alphanumeric/underscore rather than trusting
-        // an environment_variables entry that a user could have hand-edited.
-        let mut environment = test_environment();
-        environment
-            .environment_variables
-            .insert("DB_CHARSET".into(), "utf8mb4; DROP TABLE users;".into());
-        assert_eq!(database_charset(&environment), "utf8mb4");
     }
 }
